rename serialization module to blob
