+++ saffron/src/blob.rs
